# swagger-ui/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "SEMOSS",
    "version" : "0.0.1-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "https://workshop.cfg.deloitte.com/cfg-ai-demo/Monolith",
    "description" : "Demo Endpoint URL"
  }, {
    "url" : "http://localhost:8080/Monolith",
    "description" : "Local Development"
  } ],
...
  "paths" : {
...
    "/engine/runPixel" : {
      "post" : {
        "operationId" : "runPixelSync",
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "application/json;charset=utf-8" : { }
            }
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
    "/ext/mcp/{toolbox_id}/comms" : {
      "post" : {
        "operationId" : "comms",
        "parameters" : [ {
          "name" : "toolbox_id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "access_key",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "text/event-stream" : { }
            }
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
    "/ext/mcp/{toolbox_id}/it" : {
      "post" : {
        "operationId" : "getInsightData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "default" : {
            "description" : "default response",
            "content" : {
              "text/plain" : { }
            }
          }
        },
        "security" : [ {
          "basicAuth" : [ ]
        } ]
      }
    },
...
  }
}